+++ forge/src/main/java/com/jahirtrap/walljump/WallJumpClient.java
@@ -1,6 +1,10 @@
 package com.jahirtrap.walljump;
 
+import com.jahirtrap.walljump.init.ModConfig;
+import com.jahirtrap.walljump.sound.FallingSoundInstance;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.LocalPlayer;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.ClientRegistry;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -10,6 +14,8 @@ import org.lwjgl.glfw.GLFW;
 
 @Mod.EventBusSubscriber(modid = WallJumpMod.MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class WallJumpClient {
+    public static FallingSoundInstance FALLING_SOUND;
+
     public static KeyMapping KEY_WALL_JUMP = new KeyMapping(
             "key.walljump.walljump",
             GLFW.GLFW_KEY_LEFT_SHIFT,
@@ -19,5 +25,12 @@ public class WallJumpClient {
     @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         ClientRegistry.registerKeyBinding(KEY_WALL_JUMP);
+    }
+
+    public static void playFallingSound(LocalPlayer player) {
+        if (ModConfig.playFallingSound && player.fallDistance > 1.5 && !player.isFallFlying() && (FALLING_SOUND == null || FALLING_SOUND.isStopped())) {
+            FALLING_SOUND = new FallingSoundInstance(player);
+            Minecraft.getInstance().getSoundManager().play(FALLING_SOUND);
+        }
     }
 }